rename use_case ByLanguage to by_language
