+++ src/main/java/data_access/ApiSearchByRegionDataAccessObject.java
@@ -8,7 +8,7 @@ import okhttp3.Response;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import use_case.search.by_region.SearchByRegionDataAccessInterface;
-import use_case.search.ByLanguage.SearchByLanguageCountryDataAccessInterface;
+import use_case.search.by_language.SearchByLanguageCountryDataAccessInterface;
 
 import java.io.IOException;
 import java.util.*;
